+++ プロジェクトファイル/Scene/HomeScene/HomeStage.cpp
@@ -57,7 +57,7 @@ void HomeStage::Initialize()
 	Camera::SetUpDirection(XMVectorSet(0, 1, 0, 0));
 
 	//ライトの強さ
-	Light::SetPlayerIntensity(lightIntensity_);
+	Light::SetPlayerIntensity((float)lightIntensity_);
 
 	//Warpの移動先
 	warpPos_ = { 0, 15, 0 };
@@ -76,7 +76,7 @@ void HomeStage::Initialize()
 //更新
 void HomeStage::Update()
 {
-	tSpace_.rotate_.y += 0.2;
+	tSpace_.rotate_.y += 0.2f;
 }
 
 //描画
