+++ プロジェクトファイル/Scene/LastScene/LastStage.cpp
@@ -50,7 +50,7 @@ void LastStage::Initialize()
 	Camera::SetUpDirection(UP_VECTOR);
 
 	//ƒ‰ƒCƒg‚Ì‹­‚³
-	Light::SetPlayerIntensity(lightIntensity_);
+	Light::SetPlayerIntensity((float)lightIntensity_);
 
 	//Player‚Ì‰ŠúˆÊ’u
 	ARGUMENT_INITIALIZE(pos_,XMFLOAT3(-3,2,-5));
